requirements.txt
numpy
opencv-python
requests
timm
torch==1.10.2+cpu
torchvision==0.11.3+cpu
fastapi
uvicorn
Pillow
